Use variables directly instead of testing for bool
The statements `if x == True`, `if x is True` and `if x` are
functionally equivalent. Note that if you want to compare to a boolean,
the pattern `if x is True` is preferred in Python. Keep in mind to not
compare variables using `is` though, because `is` compares the identity
of variables instead of their values.
Additionally, I've replaced the second `if` condition inside your
`while` loop with a simple `else` clause. If the first if-clause
evaluates to false, the `else` clause will evaluate. Otherwise, the
first one evaluates as expected.