Remove redundant `if` after the loops
The loops in `internet_on` evaluate as long as their condition is
`False`, which means that if the loops are finished, it is guaranteed
that the condition in the loop evaluates to `False` (unless you exited
the loop manually using `break`). Thus, the `if attempts == 0`
statements after the loops are redundant.