Use for-range loop instead of while loop
This for-range loop is functionally equivalent to the `while` loop,
except that you do not manually have to decrement the variable. Instead,
this loop executes 15 times (like the while loop). The variable in place
of `_` is usually the value given by the iterator (which is `range` here),
but since we do not care about the value here we use an underscore to
signal it.
The `pass` statement in line `23` is needed because Python expects a
statement there, and we want to catch the error instead of having it
print on the console and return from the function, so we use `pass` here
to tell Python that nothing happens in that block.